Hoist transaction_info into base job class
